исправлено согласно с замечаниями
