+++ models/contacts/add-contact.js
@@ -1,0 +1,21 @@
+
+import fs from 'fs/promises';
+import path from 'path';
+import { randomUUID } from 'crypto';
+import contacts from '../../db/contacts.json';
+
+import { fileURLToPath } from 'url'; 
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+const addContact = async ({ name, email, phone }) => {
+    const newContact = { name, email, phone, id: randomUUID() };
+    contacts.push(newContact);
+    console.log('__dirname: ', __dirname)
+    await fs.writeFile(
+        path.join(__dirname, '../../db/contacts.json'),
+        JSON.stringify(contacts, null, 2)
+    );
+    return newContact;
+};
+
+export default addContact;
