+++ models/contacts/index.js
@@ -1,0 +1,8 @@
+
+import listContacts from './list-contacts';
+import getContactById from './get-contact';
+import addContact from './add-contact';
+import removeContact from './remove-contact';
+import updateContact from './update-contact';
+
+export default { listContacts, getContactById, addContact, removeContact, updateContact };
