+++ models/contacts/list-contacts.js
@@ -1,0 +1,5 @@
+import contacts from '../../db/contacts.json';
+
+const listContacts = async () => contacts;
+
+export default listContacts;
