+++ routes/api/contacts.js
@@ -1,40 +1,20 @@
 import {Router} from 'express';
 const router = new Router();
-import model from '../../model/index';
-import { validateCreate, validateUpdate, validateId } from '../api/validation';
+import { validateCreate, validateUpdate, validateId } from '../../midllewares/validation/contactsValidation';
+import getContactsController from '../controllers/contacts/get-contacts';
+import getContactController from '../controllers/contacts/get-contact';
+import deleteContactController from '../controllers/contacts/delete-contact';
+import postContactController from '../controllers/contacts/post-contact';
+import putContactController from '../controllers/contacts/put-contact';
 
-router.get('/', async (req, res, next) => {
-  const contacts = await model.listContacts();
-  res.status(200).json(contacts)
-});
+router.get('/', getContactsController);
 
-router.get('/:id', validateId, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.getContactById(id);
-  contact ?
-    res.status(200).json( contact ) :
-    res.status(404).json({ message: 'Not found'});
-});
+router.get('/:id', validateId, getContactController);
 
-router.post('/', validateCreate, async (req, res, next) => {
-  const newContact = await model.addContact(req.body);
-  res.status(201).json( newContact );
-});
+router.post('/', validateCreate, postContactController);
 
-router.delete('/:id', validateId, async (req, res, next) => {
-  const { id } = req.params
-  const deletedContact = await model.removeContact(id);
-  deletedContact ?
-    res.status(200).json({ message: 'contact deleted'}) :
-    res.status(404).json({ message: 'Not found' });
-});
+router.delete('/:id', validateId, deleteContactController);
 
-router.put('/:id', validateId, validateUpdate, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.updateContact(id, req.body);
-  contact ?
-    res.status(200).json( contact ) :
-    res.status(404).json({ message: 'Not found'});
-});
+router.put('/:id', validateId, validateUpdate, putContactController);
 
 export default router;
